requirements.txt
torch>=1.12.0
torchvision>=0.13.0
diffusers>=0.21.0
transformers>=4.21.0
accelerate>=0.20.0
tqdm>=4.64.0
wandb>=0.13.0
numpy>=1.21.0
Pillow>=9.0.0

